правки в реализации кучи

+++ src/main/java/lab/second/Node.java
@@ -1,17 +1,17 @@
 package lab.second;
 
 class Node {
-    int element, value;
+    double element, value;
     Node left, right;
 
-    public Node(int element) {
+    public Node(double element) {
         this(element, null, null);
     }
 
-    public Node(int element, Node left, Node right) {
+    public Node(double element, Node left, Node right) {
         this.element = element;
         this.left = left;
         this.right = right;
-        this.value = 0;
+        this.value = 0.0;
     }
 }
